Add tests for CORDEX domain handling in regrid functionality
- Introduced a new test to validate that non-string inputs are rejected by the CORDEX domain detection.
- Updated existing tests to ensure proper handling of CORDEX domains in target grid construction and validation.
- Removed redundant tests that were previously validating CORDEX domain acceptance in the update target grid function.
